Move Mattermost to Bot
